+++ tests/lib/rules/jsx-fragments.js
@@ -141,6 +141,19 @@ ruleTester.run('jsx-fragments', rule, {
       ],
     },
     {
+      code: '<><Foo /></>',
+      output: '<><Foo /></>', // should get '<Act.Frag><Foo /></Act.Frag>', but the old TS parser lacks opening/closing Fragment info
+      features: ['fragment', 'no-babel', 'ts', 'no-ts-new'],
+      options: ['element'],
+      settings,
+      errors: [
+        {
+          messageId: 'preferPragma',
+          data: { react: 'Act', fragment: 'Frag' },
+        },
+      ],
+    },
+    {
       code: '<Act.Frag><Foo /></Act.Frag>',
       output: '<><Foo /></>',
       options: ['syntax'],